// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  USER: 'USER',
  EDITOR: 'EDITOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ImgStatus = {
  TEMP: 'TEMP',
  USED: 'USED'
} as const

export type ImgStatus = (typeof ImgStatus)[keyof typeof ImgStatus]


export const Imagetemptype = {
  COVER: 'COVER',
  BANNER: 'BANNER'
} as const

export type Imagetemptype = (typeof Imagetemptype)[keyof typeof Imagetemptype]